requirements.txt
Pillow
watchdog
imageio
numpy
pyautogui
opencv-python
ffmpeg-python
googletrans